Changed position to double for extra precision.

+++ include/position.h
@@ -7,12 +7,12 @@ namespace StratusQuo
     {
         public: 
             Position();
-            float getX();
-            float getY();
-            void changeX(float delta);
-            void changeY(float delta);
+            double getX();
+            double getY();
+            void changeX(double delta);
+            void changeY(double delta);
         private:
-            float x = 0, y = 0;
+            double x = 0, y = 0;
     };
 } // namespace StratusQuo
 #endif
